[feature] Rename to HttpsListener and Module-ize
.
