--- test/tools/ClientCertificateCheck/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using ClientCertificateCheck
-@namespace ClientCertificateCheck.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
