+++ test/tools/HttpsListener/HttpsListener.csproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <Description>Very simple ASP.NET Core 2.0 app to provide client certificate details if one is supplied in the webrequest.</Description>
+    <Description>Very simple ASP.NET Core 2.0 app to provide an HTTPS server.</Description>
     <TargetFramework>netcoreapp2.0</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
