+++ test/tools/HttpsListener/Pages/Error.cshtml.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace ClientCertificateCheck.Pages
+namespace HttpsListener.Pages
 {
     public class ErrorModel : PageModel
     {
