+++ test/tools/HttpsListener/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using HttpsListener
+@namespace HttpsListener.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
