+++ test/tools/HttpsListener/Startup.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ClientCertificateCheck
+namespace HttpsListener
 {
     public class Startup
     {
